Migrate test_mc_specphot_kern_merging to test_specphot_kernels_merging.py
